requirements.txt
openai
requests
beautifulsoup4
python-telegram-bot
apscheduler